Fix Pinecone API camelCase serialization for delete and query operations
The Pinecone REST API expects camelCase field names in JSON payloads (deleteAll, topK, includeMetadata), but the Rust code was using snake_case which Serde serialized as-is. This caused the delete operation to fail with a 404 error. Added #[serde(rename_all = "camelCase")] to DeleteRequest and QueryRequest structs to fix the serialization.